+++ shop/WEB-INF/src/controller/cart/add.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package controller.cart;
 
 import java.io.IOException;
 
@@ -8,18 +8,18 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-@WebServlet("/cart/index")
-public class index extends HttpServlet {
+@WebServlet("/cart/add")
+public class add extends HttpServlet {
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response)
   throws IOException, ServletException {
     response.setContentType("text/html; charset=UTF-8");
     request.setCharacterEncoding("UTF-8");
 
-    // 処理
+    String id = request.getParameter("id");
 
-    getServletContext()
-      .getRequestDispatcher("/resources/jsp/cart/index.jsp")
-        .forward(request, response);
+    System.out.println("商品ID: " + id);
+
+    response.sendRedirect("index");
   }
 }
