+++ shop/WEB-INF/src/controller/cart/index.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package controller.cart;
 
 import java.io.IOException;
 
@@ -8,18 +8,18 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-@WebServlet("/cart/add")
-public class add extends HttpServlet {
+@WebServlet("/cart/index")
+public class index extends HttpServlet {
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response)
   throws IOException, ServletException {
     response.setContentType("text/html; charset=UTF-8");
     request.setCharacterEncoding("UTF-8");
 
-    String id = request.getParameter("id");
+    // 処理
 
-    System.out.println("商品ID: " + id);
-
-    response.sendRedirect("index");
+    getServletContext()
+      .getRequestDispatcher("/resources/jsp/cart/index.jsp")
+        .forward(request, response);
   }
 }
