+++ shop/WEB-INF/src/controller/cashresister/create.java
@@ -1,4 +1,4 @@
-package cashresister.controller;
+package controller.cashresister;
 
 import java.io.IOException;
